shinpj/movietts/backend/main.py
import os

from dotenv import load_dotenv

load_dotenv()

from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from backend.routers.subtitle import UPLOAD_DIR, router as subtitle_router

app = FastAPI(
    title="MovieTTS API",
    description="영상 음성 분석 및 자막 자동 생성 API",
    version="0.1.0",
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:5173"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(subtitle_router)

os.makedirs(UPLOAD_DIR, exist_ok=True)


@app.get("/")
async def root():
    return {"message": "MovieTTS API", "version": "0.1.0"}
