shinpj/movietts/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai-whisper>=20231117
python-multipart>=0.0.6
pydantic>=2.0
aiofiles>=23.0
yt-dlp>=2024.0
google-genai>=1.0.0
